+++ #client/prc_cli_manualscan.py
@@ -1,4 +1,6 @@
 # prc_template  v 0.2.0
+import sys
+sys.path.append("..")
 import time
 import datetime
 import pygame
@@ -10,8 +12,8 @@ from fRedis import clsRedis
 from pygame import mixer
 
 def start_process(config_file,__cli_id__):
-    __prc_cli_type__=f"cli_playsound"
-    __prc_name__=f"cli%02d_playsound"%(__cli_id__,)
+    __prc_cli_type__=f"cli_manualscan"
+    __prc_name__=f"cli%02d_manualscan"%(__cli_id__,)
     
     ini_config = clsConfig(config_file)   # 来自主线程的配置文件
     inst_logger = clsLogger(ini_config)  
@@ -30,24 +32,32 @@ def start_process(config_file,__cli_id__):
     __prc_cycletime=__ini_prc_config__.CycleTime.prc_cycletime
     __prc_expiretime=__ini_prc_config__.CycleTime.prc_expiretime
     __prc_healthytime=__ini_prc_config__.CycleTime.prc_healthytime
-    
+
     # 取得ini文件中的全部声音资源列表
     # ToDo 自动遍历ini文件中，Sound字段下的配置文件
     dict_sound = {}
-    dict_sound['reading_gr']= __ini_prc_config__.Sound.reading_gr
-    dict_sound['reading_nr']= __ini_prc_config__.Sound.reading_nr
-    dict_sound['reading_mr']= __ini_prc_config__.Sound.reading_mr
+    dict_sound['ms_barcode_reject']= __ini_prc_config__.Sound.ms_barcode_reject
+    dict_sound['ms_barcode_exist']= __ini_prc_config__.Sound.ms_barcode_exist
+    dict_sound['ms_barcode_rescan_accpet']= __ini_prc_config__.Sound.ms_barcode_rescan_accpet    
+
+    # 取得ini文件中的全部条码正则表达式列表
+    # ToDo 自动遍历ini文件中，barcode字段下的所有正则表达式
+    lst_re_exp = []
+    lst_re_exp.append(__ini_prc_config__.Barcode.re_exp_01)
+    lst_re_exp.append(__ini_prc_config__.Barcode.re_exp_02)
     
     # 增加Redis中总线程计数器，并将增加后的计数器值作为当前线程的id
     __prc_id__ = inst_redis.incrkey(f"pro_mon:prc_counter")
     
     inst_logger.info("线程 %s 取得 prc_id = %d, cli_id = %d"%(__prc_name__,__prc_id__,__cli_id__)) 
-    inst_redis.setkeypx(f"pro_mon:{__prc_name__}:run_lock",__prc_id__,__prc_expiretime)
-    inst_logger.info("线程 %s 已设置线程锁，过期时间 = %d ms"%(__prc_name__,__prc_expiretime)) 
+    # inst_redis.setkeypx(f"pro_mon:{__prc_name__}:run_lock",__prc_id__,__prc_expiretime)
+    inst_redis.setkey(f"pro_mon:{__prc_name__}:run_lock",__prc_id__)             ## manucal scan函数不允许超时 ##      
+    #inst_logger.info("线程 %s 已设置线程锁，过期时间 = %d ms"%(__prc_name__,__prc_expiretime)) 
+    inst_logger.info("线程 %s 已设置线程锁，过期时间 = %d ms"%(__prc_name__,-1))      ## manucal scan函数不允许超时 ##
 
     # 增加当前线程的重启次数,如为1说明是首次启动
-    __prc_restart__ = inst_redis.incrkey(f"pro_mon:{__prc_name__}:restart")           ## manucal scan函数不记录重启 ##
-    inst_logger.info("线程 %s 启动次数 restart = %d"%(__prc_name__,__prc_restart__))    ## manucal scan函数不记录重启 ##
+    # __prc_restart__ = inst_redis.incrkey(f"pro_mon:{__prc_name__}:restart")           ## manucal scan函数不记录重启 ##
+    # inst_logger.info("线程 %s 启动次数 restart = %d"%(__prc_name__,__prc_restart__))    ## manucal scan函数不记录重启 ##
         
     # 记录线程启动时间
     __prc_start_ts__ = datetime.datetime.now()
@@ -64,44 +74,58 @@ def start_process(config_file,__cli_id__):
    
     b_thread_running = True
     int_exit_code = 0
-    
+    lst_reading_gr =[]
+    lst_reading_nr =[]
+    lst_reading_mr =[]
+
     # 初始化pygame.mixer
     pygame.mixer.init()
     
-    if not inst_redis.xcreategroup("stream_test", __prc_name__):
-        inst_logger.info("线程 %s 注册stream组失败，该组已存在" %("__prc_name__",))
-    else:
-        inst_logger.info("线程 %s 注册stream组成功" %("__prc_name__",))
-    
     while b_thread_running:
         # 刷新当前线程的运行锁
-        inst_redis.setkeypx(f"pro_mon:{__prc_name__}:run_lock",__prc_id__,__prc_expiretime)
+        # inst_redis.setkeypx(f"pro_mon:{__prc_name__}:run_lock",__prc_id__,__prc_expiretime)
+        # inst_redis.setkey(f"pro_mon:{__prc_name__}:run_lock",__prc_id__)             ## manucal scan函数不允许超时     
         
         # --------------------
         # 主线程操作区
+        strManualScanBarcode = input("please enter manual scan barcode...")
         
-        lst_reading_nr = list(inst_redis.getset("set_reading_nr"))  # 更新set_reading_nr
-        set_reading_mr = inst_redis.getset("set_reading_mr")        # 更新set_reading_mr
-        l = inst_redis.xreadgroup("stream_test",__prc_name__,"cliplay-id01")
-        if len(lst_reading_nr) + len(set_reading_mr) == 0:          # 正常状态，读取 stream_test, 每收到一个gr 播放一个声音l
-            if len(l)>0:                                # 收到消息
-                inst_logger.info("收到序列 %s 中的消息累计 %d 行" %(l[0][0],len(l[0][1])))
-                for i,dictdata in l[0][1]:              # 遍历收到的所有消息
-                    if dictdata['result']=='GR':        # 正常识读
-                        mixer.music.load(dict_sound['reading_gr'])
-                        mixer.music.play()
-                    elif dictdata['result']=='MR':      # 多条码
-                        mixer.music.load(dict_sound['reading_mr'])
-                        mixer.music.play()
-                    elif dictdata['result']=='NR':      # 无条码    
-                        mixer.music.load(dict_sound['reading_nr'])
-                        mixer.music.play()                        
-        else:                                                       # 补码状态，收取HIK的读码信息，但是不播放声音，只播放补码声音
-            time.sleep(__prc_cycletime/1000.0)  # 所有时间均以ms形式存储
-            
+        # 更新set_reading_gr
+        lst_reading_gr = inst_redis.getset("set_reading_gr")
+        # 更新set_reading_nr
+        # lst_reading_nr = inst_redis.getset("set_reading_nr")
+        # 更新set_reading_mr
+        lst_reading_mr = inst_redis.getset("set_reading_mr")         
 
+        if strManualScanBarcode in lst_reading_gr:
+            print("Barcode is already exist!!")
+            mixer.music.load(dict_sound['ms_barcode_exist'])
+            mixer.music.play()
+            continue
+        
+        if strManualScanBarcode in lst_reading_mr:      
+            print("Get MRead Barcode !!")
+            inst_redis.xadd( "stream_manualscan", {'cli_id':__cli_id__,'scan_id':__prc_id__,'barcode':strManualScanBarcode,'type':'MR'})      # 插入 Manual Scan stream/MR
+            mixer.music.load(dict_sound['ms_barcode_rescan_accpet'])
+            mixer.music.play()
+            continue
+        
+        # 条码格式校验
+        bBarcodeValid = False
+        for i, re_exp in enumerate(lst_re_exp):
+            if barcode_formatcheck(strManualScanBarcode,re_exp):
+                inst_redis.xadd( "stream_manualscan", {'cli_id':__cli_id__,'scan_id':__prc_id__,'barcode':strManualScanBarcode,'type':'NR'})      # 插入 Manual Scan stream/NR
+                print("Barcode valid,insert into system")
+                bBarcodeValid = True
+                mixer.music.load(dict_sound['ms_barcode_rescan_accpet'])
+                mixer.music.play()                
+                break
+        if not bBarcodeValid:
+            print("Barcode is not valid!!")
+            mixer.music.load(dict_sound['ms_barcode_reject'])
+            mixer.music.play()
         # --------------------
-        # time.sleep(__prc_cycletime/1000.0)  # 所有时间均以ms形式存储
+        time.sleep(__prc_cycletime/1000.0)  # 所有时间均以ms形式存储
         
         # 线程运行时间与健康程度判断